// frontend/src/app/models/rezervacija.ts
export class Rezervacija{
  idR: number= 0;
  datumZakazivanja: string = '';
  vreme:  string = '';
  cena: number= 0;
  kozmeticar: string = '';
  klijent: string = '';
  napomena: string = '';
  status: string = '';
  idU: number= 0;
  idPU: number= 0;
  nazivPodusluge: string = "";
  komentar: string = '';
}